requirements.txt
python-telegram-bot>=20.0
google-generativeai>=0.8.0
groq>=0.9.0
python-dotenv>=1.0.0
python-docx>=1.1.0
Pillow>=10.0.0
pymupdf>=1.24.0
libsql-client>=0.3.0
